ui: Refactor Confirmationdialog
If an intent is made by the user and a saga function is called the saga triggers a CONFIRM_INTENT action and waits until INTENT_CONFIRM or INTENT_CANCELED is triggered.
The ConfirmationContainers reducer saves the intent (e.g. project.assign) and the payload/data that is necessary to:
 - Check if assignee has view permissions
 - Check if assigner is permitted to grant these permissions
 - Build individual dialog content depending on the intent
When the decision is made by the user only a INTENT_CONFIRM or INTENT_CANCELED action is triggered and the saga continue finishing or cancel its action.